Read the insertion parameter table in double
The value-conversion table every insertion effect resolves its
parameters through was a float[128][14]; upstream's is number[][], so
each entry came back rounded to float32.  Its rate column feeds an LFO
phase accumulator, which is where that kind of error compounds.

The phaser's own setters had the same shape as the chorus ones before
them -- computed in float, stored into double fields -- so those and its
reset defaults are widened too.

phaser 78.81 -> 80.45 dB.  The rest of that file's gap is not in the
phaser: every constant, coefficient and expression now matches upstream,
and phaser_manual_param, which drives the same effect but sweeps only
the manual parameter, sits at 137.41 dB.  What separates them is that
phaser also sweeps resonance, and eight all-pass stages with feedback
amplify whatever difference arrives at the input.

+++ spessasynth_core/src/synthesizer/audio_engine/effects/insertion/convert.c
@@ -9,7 +9,7 @@
 
 /* ── Conversion table (InsertionValueConverter) ───────────────────────────── */
 /* Columns: PreDly,Dly1,Dly2,Dly3,Dly4, Rate1,Rate2, HF,Cut,EQ,LPF,Manual,Azim,Accl */
-static const float IVC[128][14] = {
+static const double IVC[128][14] = {
 	{ 0.0f, 200, 200, 0.0f, 0, 0.05f, 0.05f, 315, 250, 200, 250, 100, -180, 0 },
 	{ 0.1f, 205, 205, 0.1f, 5, 0.10f, 0.10f, 315, 250, 200, 250, 110, -180, 1 },
 	{ 0.2f, 210, 210, 0.2f, 10, 0.15f, 0.15f, 315, 250, 200, 250, 120, -180, 2 },
@@ -140,12 +140,12 @@ static const float IVC[128][14] = {
 	{ 100.f, 1000, 1000, 500.f, 635, 10.0f, 6.40f, 13500, 8000, 6300, 13500, 8000, -180, 15 },
 };
 
-float ss_ivc_rate1(int v) {
+double ss_ivc_rate1(int v) {
 	return IVC[v & 127][5];
 }
-float ss_ivc_manual(int v) {
+double ss_ivc_manual(int v) {
 	return IVC[v & 127][11];
 }
-float ss_ivc_eq_freq(int v) {
+double ss_ivc_eq_freq(int v) {
 	return IVC[v & 127][9];
 }
